starting to build parser
